requirements.txt
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloud-init==23.2.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
DateTime==5.4
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.9
djangorestframework==3.15.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.12.3
Fiona==1.9.4.post1
frozenlist==1.4.0
geopandas==0.13.2
gpg==1.13.1
grpcio==1.57.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.0.1
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==1.1.0
more-itertools==4.2.0
msgpack==1.0.5
netifaces==0.10.4
oauthlib==3.1.0
openpyxl==3.1.2
optional-django==0.1.0
packaging==23.1
pandas==2.0.1
Parsley==1.3
pdoc3==0.10.0
pexpect==4.6.0
phonenumbers==8.13.28
pkgutil-resolve-name==1.3.10
protobuf==4.24.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyproj==3.5.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
pytz==2023.3
PyYAML==6.0.1
ray==2.6.3
referencing==0.30.2
requests==2.31.0
requests-unixsocket==0.2.0
rpds-py==0.10.2
SecretStorage==2.3.1
service-identity==18.1.0
sgp4==2.22
shapely==2.0.1
simplejson==3.16.0
six==1.14.0
sos==4.5.6
sqlparse==0.4.4
ssh-import-id==5.10
systemd-python==234
tabulate==0.9.0
Twisted==18.9.0
typing-extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==2.0.4
wadllib==1.3.3
zipp==3.16.2
zope.interface==4.7.1
